chore(tracer): documentation and overflow-guard nits
M1 (modal_state.rs content_modal_scroll_to): replace terse "no fire here"
comment with a fuller note explaining the DIFF_SIZE_CAP_BYTES cap and
eager loading. [Applied in Fix 5 refactor commit.]

M3 (modal_state.rs open_content_modal): add comment above last_nondiff_tab
assignment explaining its dual role (Save target + Tab cycling default).
[Applied in Fix 5 refactor commit.]

M5 (modal.rs delta_ui): use i128 subtraction instead of i64 to eliminate
the theoretical overflow on extreme u64 byte counts. unsigned_abs() then
returns u128, which as f64 is safe for display purposes.

+++ src/view/tracer/modal.rs
@@ -341,7 +341,7 @@ fn size_ui(n: Option<u64>) -> String {
 fn delta_ui(i: Option<u64>, o: Option<u64>) -> String {
     match (i, o) {
         (Some(i), Some(o)) => {
-            let d = o as i64 - i as i64;
+            let d = o as i128 - i as i128;
             let sign = if d >= 0 { "+" } else { "-" };
             let abs = d.unsigned_abs();
             if abs >= 1024 {
